+++ crates/prql_compiler/src/codegen/mod.rs
@@ -1,0 +1,312 @@
+use prql_ast::expr::*;
+use prql_ast::stmt::*;
+
+mod literal;
+mod ast;
+mod types;
+pub use literal::DisplayLiteral;
+
+pub fn write_stmts(stmts: &Vec<Stmt>) -> String {
+    let mut opt = WriteOpt::default();
+
+    loop {
+        if let Some(s) = stmts.write(opt.clone()) {
+            break s;
+        } else {
+            opt.max_width += opt.max_width / 2;
+        }
+    }
+}
+
+pub fn write_expr(expr: &Expr) -> String {
+    let opt = WriteOpt::new_width(u16::MAX);
+    expr.write(opt).unwrap()
+}
+
+pub trait WriteSource {
+    /// Converts self to its source representation according to specified
+    /// options.
+    ///
+    /// Returns `None` if source does not fit into [WriteOpt::rem_width].
+    fn write(&self, opt: WriteOpt) -> Option<String>;
+
+    fn write_between<S: ToString>(
+        &self,
+        prefix: S,
+        suffix: &str,
+        mut opt: WriteOpt,
+    ) -> Option<String> {
+        let mut r = String::new();
+        r += opt.consume(&prefix.to_string())?;
+        opt.context_strength = 0;
+        opt.unbound_expr = false;
+
+        let source = self.write(opt.clone())?;
+        r += opt.consume(&source)?;
+
+        r += opt.consume(suffix)?;
+        Some(r)
+    }
+}
+
+impl<T: WriteSource> WriteSource for &T {
+    fn write(&self, opt: WriteOpt) -> Option<String> {
+        (*self).write(opt)
+    }
+}
+
+#[derive(Clone)]
+pub struct WriteOpt {
+    /// String to emit as one indentation level
+    pub tab: &'static str,
+
+    /// Maximum number of characters per line
+    pub max_width: u16,
+
+    /// Current indent used when emitting lines
+    pub indent: u16,
+
+    /// Current remaining number of characters in line
+    pub rem_width: u16,
+
+    /// Strength of the context
+    /// For top-level exprs or exprs in parenthesis, this will be 0.
+    /// For exprs in function calls, this will be 10.
+    pub context_strength: u8,
+
+    /// True iff preceding source ends in an expression that could
+    /// be mistakenly bound into a binary op by appending an unary op.
+    ///
+    /// For example:
+    /// `join foo` has an unbound expr, since `join foo ==bar` produced a binary op.
+    pub unbound_expr: bool,
+}
+
+impl Default for WriteOpt {
+    fn default() -> Self {
+        Self {
+            tab: "  ",
+            max_width: 50,
+
+            indent: 0,
+            rem_width: 50,
+            context_strength: 0,
+            unbound_expr: false,
+        }
+    }
+}
+
+impl WriteOpt {
+    fn new_width(max_width: u16) -> Self {
+        WriteOpt {
+            max_width,
+            rem_width: max_width,
+            ..WriteOpt::default()
+        }
+    }
+
+    fn consume_width(&mut self, width: u16) -> Option<()> {
+        self.rem_width = self.rem_width.checked_sub(width)?;
+        Some(())
+    }
+
+    fn reset_line(&mut self) -> Option<()> {
+        let ident = self.tab.len() as u16 * self.indent;
+        self.rem_width = self.max_width.checked_sub(ident)?;
+        Some(())
+    }
+
+    fn consume<'a>(&mut self, source: &'a str) -> Option<&'a str> {
+        let width = if let Some(new_line) = source.rfind('\n') {
+            source.len() - new_line
+        } else {
+            source.len()
+        };
+        self.consume_width(width as u16);
+        Some(source)
+    }
+
+    fn write_indent(&self) -> String {
+        self.tab.repeat(self.indent as usize)
+    }
+}
+
+struct SeparatedExprs<'a, T: WriteSource> {
+    exprs: &'a [T],
+    inline: &'static str,
+    line_end: &'static str,
+}
+
+impl<'a, T: WriteSource> WriteSource for SeparatedExprs<'a, T> {
+    fn write(&self, mut opt: WriteOpt) -> Option<String> {
+        // try inline
+        if let Some(inline) = self.write_inline(opt.clone()) {
+            return Some(inline);
+        }
+
+        // one per line
+        {
+            opt.indent += 1;
+
+            let mut r = String::new();
+
+            for expr in self.exprs {
+                r += "\n";
+                r += &opt.write_indent();
+                opt.reset_line()?;
+                opt.rem_width.checked_sub(self.line_end.len() as u16)?;
+
+                r += &expr.write(opt.clone())?;
+                r += self.line_end;
+            }
+            opt.indent -= 1;
+            r += "\n";
+            r += &opt.write_indent();
+
+            Some(r)
+        }
+    }
+}
+
+impl<'a, T: WriteSource> SeparatedExprs<'a, T> {
+    fn write_inline(&self, mut opt: WriteOpt) -> Option<String> {
+        let mut exprs = Vec::new();
+        for expr in self.exprs {
+            let expr = expr.write(opt.clone())?;
+
+            if expr.contains('\n') {
+                return None;
+            }
+            opt.consume_width(expr.len() as u16)?;
+
+            exprs.push(expr);
+        }
+
+        let separators = self.inline.len() * (exprs.len().checked_sub(1).unwrap_or_default());
+        opt.consume_width(separators as u16)?;
+
+        Some(exprs.join(self.inline))
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use insta::assert_snapshot;
+    use similar_asserts::assert_eq;
+
+    use super::*;
+
+    fn assert_is_formatted(input: &str) {
+        let stmt = format_single_stmt(input);
+
+        assert_eq!(input.trim(), stmt.trim());
+    }
+
+    fn format_single_stmt(query: &str) -> String {
+        use itertools::Itertools;
+        let stmt = crate::prql_to_pl(query)
+            .unwrap()
+            .into_iter()
+            .exactly_one()
+            .unwrap();
+        stmt.write(WriteOpt::default()).unwrap()
+    }
+
+    #[test]
+    fn test_pipeline() {
+        let short = Expr::new(ExprKind::Ident(Ident::from_path(vec!["short"])));
+        let long = Expr::new(ExprKind::Ident(Ident::from_path(vec![
+            "some_module",
+            "submodule",
+            "a_really_long_name",
+        ])));
+
+        let mut opt = WriteOpt {
+            indent: 1,
+            ..Default::default()
+        };
+
+        // short pipelines should be inlined
+        let pipeline = Expr::new(ExprKind::Pipeline(Pipeline {
+            exprs: vec![short.clone(), short.clone(), short.clone()],
+        }));
+        assert_snapshot!(pipeline.write(opt.clone()).unwrap(), @"(short | short | short)");
+
+        // long pipelines should be indented
+        let pipeline = Expr::new(ExprKind::Pipeline(Pipeline {
+            exprs: vec![short.clone(), long.clone(), long, short.clone()],
+        }));
+        // colons are a workaround to avoid trimming
+        assert_snapshot!(pipeline.write(opt.clone()).unwrap(), @r###"
+        (
+            short
+            some_module.submodule.a_really_long_name
+            some_module.submodule.a_really_long_name
+            short
+          )
+        "###);
+
+        // sometimes, there is just not enough space
+        opt.rem_width = 4;
+        opt.indent = 100;
+        let pipeline = Expr::new(ExprKind::Pipeline(Pipeline { exprs: vec![short] }));
+        assert!(pipeline.write(opt).is_none());
+    }
+
+    #[test]
+    fn test_escaped_string() {
+        assert_is_formatted(r#"filter name ~= "\\(I Can't Help\\) Falling""#);
+    }
+
+    #[test]
+    fn test_double_braces() {
+        assert_is_formatted(
+            r#"let has_valid_title = s"regexp_contains(title, '([a-z0-9]*-){{2,}}')""#,
+        );
+    }
+
+    #[test]
+    fn test_unary() {
+        assert_is_formatted(r#"sort {-duration}"#);
+
+        assert_is_formatted(r#"select a = -b"#);
+        assert_is_formatted(r#"join `project-bar.dataset.table` (==col_bax)"#)
+    }
+
+    #[test]
+    fn test_simple() {
+        assert_is_formatted(r#"aggregate average_country_salary = (average salary)"#);
+    }
+
+    #[test]
+    fn test_assign() {
+        assert_is_formatted(
+            r#"
+group {title, country} (aggregate {
+  average salary,
+  average gross_salary,
+  sum salary,
+  sum gross_salary,
+  average gross_cost,
+  sum_gross_cost = sum gross_cost,
+  ct = count salary,
+})"#,
+        );
+    }
+    #[test]
+    fn test_range() {
+        assert_is_formatted(
+            r#"
+from foo
+is_negative = -100..0
+"#,
+        );
+
+        assert_is_formatted(
+            r#"
+from foo
+is_negative = -(100..0)
+"#,
+        );
+    }
+}
